Eastern and Central Pacific HURDAT

HURDAT dataset for the northeastern and central Pacific basin.

EP

Format

A dataframe with 21 variables.

Key

Unique key identifying the tropical cyclone. Formatted like AABBCCCC where AA is Basin, BB is YearNum and CC is Year

Name

Name of the storm (UNNAMED for early storms)

DateTime

Date and time of observation in UTC, POSIXct

Record

Record identifier:

C

Closest approach to a coast, not followed by a landfall.

G

Genesis.

I

An intensity peak in terms of both pressure and wind.

L

Landfall (center of system crossing a coastline).

P

Minimum central pressure.

R

Provides additional detail on the intensity of the cyclone when rapid changes are underway.

S

Change of status of the system.

T

Provides additional detail on the track.

W

Maximum sustained wind speed.

Status

Status of the system:

TD

Tropical depression (winds < 34 kts)

TS

Tropical storm (winds 34-63 kts)

HU

Hurricane (winds >= 64 kts)

EX

Extratropical cyclone of any intensity

SD

Subtropical cyclone of subtropical depression intensity (winds < 34 kts).

SS

Subtroipcal cyclone of subtropical cyclone intensity (winds >= 34 kts)

LO

A low that is neither a tropical cyclone, subtropical cyclone nor extratropical of any intensity.

WV

Tropical wave

DB

Tropical disturbance

Lat

Latitude position; >0 for northern hemisphere, <0 for southern hemisphere.

Lon

Longitude position; >0 for eastern hemisphere, <0 for western hemisphere.

Wind

Recorded or estimated wind speed in knots

Pressure

Recorded or estimated central pressure in millibars

NE34

Radius of winds > 34 kts in the northeastern quadrant.

SE34

Radius of winds > 34 kts in the southeastern quadrant.

SW34

Radius of winds > 34 kts in the southwestern quadrant.

NW34

Radius of winds > 34 kts in the northwestern quadrant.

NE50

Radius of winds > 50 kts in the northeastern quadrant.

SE50

Radius of winds > 50 kts in the southeastern quadrant.

SW50

Radius of winds > 50 kts in the southwestern quadrant.

NW50

Radius of winds > 50 kts in the northwestern quadrant.

NE64

Radius of winds > 64 kts in the northeastern quadrant.

SE64

Radius of winds > 64 kts in the southeastern quadrant.

SW64

Radius of winds > 64 kts in the southwestern quadrant.

NW64

Radius of winds > 64 kts in the northwestern quadrant.